+++ src/merge/track_info.h
@@ -299,8 +299,8 @@ public:
   std::map<int64_t, uint64_t> m_field_order_list; // As given on the command line
   option_with_source_c<uint64_t> m_field_order;   // For this very track
 
-  std::map<int64_t, int> m_colour_matrix_list; // As given on the command line
-  option_with_source_c<int> m_colour_matrix; // For this very track
+  std::map<int64_t, int> m_colour_matrix_coeff_list; // As given on the command line
+  option_with_source_c<int> m_colour_matrix_coeff; // For this very track
 
   std::map<int64_t, int> m_bits_per_channel_list; // As given on the command line
   option_with_source_c<int> m_bits_per_channel; // For this very track